+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This changelog only contains the changes that are unreleased. For changes for in
 ### New Features
 
 - Support installing datapacks to the datapacks folder and support for adding and managing them
+- Add Modrinth download metadata to Modrinth CDN downloads [#1028]
 
 ### Fixes